allow puppeteer client to be dependency injected and stub in tests
