+++ packages/preview-service/src/server/server.ts
@@ -5,42 +5,19 @@ import { getAppPort, getHost, getMetricsHost, getMetricsPort } from '@/utils/env
 import http from 'http'
 import type { Knex } from 'knex'
 import { isNaN, isString, toNumber } from 'lodash-es'
-import { PuppeteerClient, puppeteerClientFactory } from '@/clients/puppeteer.js'
-import { extendLoggerComponent, logger } from '@/observability/logging.js'
-import { puppeteerDriver } from '@/scripts/puppeteerDriver.js'
-import {
-  getChromiumExecutablePath,
-  getPreviewTimeout,
-  getPuppeteerUserDataDir,
-  serviceOrigin,
-  shouldBeHeadless
-} from '@/utils/env.js'
+import type { PuppeteerClient } from '@/clients/puppeteer.js'
 
-export const startServer = async (params: {
+export const startServer = (params: {
   db: Knex
   serveOnRandomPort?: boolean
+  puppeteerClient: PuppeteerClient
 }) => {
-  const { db } = params
+  const { db, serveOnRandomPort, puppeteerClient } = params
 
-  const puppeteerClient = await puppeteerClientFactory({
-    logger: extendLoggerComponent(logger, 'puppeteerClient'),
-    url: `${serviceOrigin()}/render/`,
-    script: puppeteerDriver,
-    launchParams: {
-      headless: shouldBeHeadless(),
-      userDataDir: getPuppeteerUserDataDir(),
-      executablePath: getChromiumExecutablePath(),
-      protocolTimeout: getPreviewTimeout(),
-      // we trust the web content that is running, so can disable the sandbox
-      // disabling the sandbox allows us to run the docker image without linux kernel privileges
-      args: ['--no-sandbox', '--disable-setuid-sandbox', '--disable-dev-shm-usage']
-    },
-    timeoutMilliseconds: getPreviewTimeout()
-  })
   /**
    * Get port from environment and store in Express.
    */
-  const inputPort = params.serveOnRandomPort ? 0 : normalizePort(getAppPort())
+  const inputPort = serveOnRandomPort ? 0 : normalizePort(getAppPort())
   const app = appFactory({ db, puppeteerClient })
   app.set('port', inputPort)
 
@@ -77,7 +54,7 @@ export const startServer = async (params: {
   })
   metricsServer.listen(inputMetricsPort, metricsHost)
 
-  return { app, server, metricsServer, puppeteerClient }
+  return { app, server, metricsServer }
 }
 
 export const stopServer = async (params: {
